backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.11.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.14.0
anthropic==0.18.1
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-json-logger==2.0.7
